将QuickLZ隔离入namespace并编写压缩wrapper
在Core i5-5250U上用编译参数
clang++ -O2 -march=native -mtune=native
可以得到409MB/s左右的压缩速度

+++ shared/fundamental_structure.h
@@ -23,8 +23,8 @@ typedef char s8;
 typedef unsigned char u8;
 typedef short s16;
 typedef unsigned short u16;
-typedef long s32;
-typedef unsigned long u32;
+typedef int s32;
+typedef unsigned int u32;
 typedef long long s64;
 typedef unsigned long long u64;
